Refactor Table.put_item tests
