+++ tests/test_put_item.py
@@ -1,0 +1,43 @@
+from dataclasses import dataclass
+
+from amano import Attribute, Item, Table
+
+
+def test_can_put_item(default_dynamodb_client, default_table) -> None:
+    # given
+    @dataclass
+    class Track(Item):
+        artist_name: str
+        track_name: str
+        album_name: str
+
+    my_table = Table[Track](default_dynamodb_client, default_table)
+
+    # when
+    item = Track("Tool", "Reflection", "Lateralus")
+    result = my_table.put(item)
+
+    # then
+    assert result
+
+
+def test_can_put_item_with_condition(
+    default_dynamodb_client, default_table
+) -> None:
+
+    # given
+    @dataclass
+    class Track(Item):
+        artist_name: Attribute[str]
+        track_name: Attribute[str]
+        album_name: Attribute[str]
+
+    my_table = Table[Track](default_dynamodb_client, default_table)
+
+    # when
+    item = Track("Tool", "Reflection", "Lateralus")
+    result = my_table.put(item, condition=Track.artist_name.not_exists())
+
+    # then
+    assert result
+
